Django/0315/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-markdown-deux==1.0.6
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
markdown2==2.4.13
packaging==24.0
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
